Rename ProfileScreen.tsx -> ProfilesScreen.tsx to match component name
